Helper functions and comments
Change initialization vector variable name
Change helpers.js to helper_functions.js
