--- src/helpers.js
@@ -1,5 +1,0 @@
-function _d(__) { console.log(__); return __; }
-
-export {
-  _d
-}
